pyproject.toml
[tool.poetry]
name = "ai-research"
version = "1.3.0"
description = "AI-native research workflow with multi-provider LLM peer review"
authors = ["Jazz <jazz@example.com>"]
readme = "README.md"
packages = [{include = "research_cli"}]

[tool.poetry.dependencies]
python = "^3.10"
anthropic = "^0.39.0"
openai = "^1.58.0"
click = "^8.1.7"
pydantic = "^2.10.0"
python-dotenv = "^1.0.0"
rich = "^13.9.0"
aiohttp = "^3.11.0"
pyyaml = "^6.0.2"
fastapi = ">=0.115.0"
uvicorn = {version = ">=0.32.0", extras = ["standard"]}
gunicorn = ">=22.0.0"
google-genai = ">=1.0.0"

[tool.poetry.scripts]
ai-research = "research_cli.cli:cli"

[tool.pytest.ini_options]
markers = [
    "llm: tests that make real LLM API calls (may be slow and require API keys)",
]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
